+++ frontend/src/pages/QueryDocuments/stores/useDocTasksCurrent.ts
@@ -36,6 +36,8 @@ interface DocTasksCurrentState {
   setNeedReload: (needReload: boolean) => void;
   setFromHistory: (item: DocTasksShortItem) => void;
   setNewQuery: () => void;
+  previousQuery: DocTasksQuery | null;
+  isSameAsPreviousQuery: (query: DocTasksQuery) => boolean;
 }
 
 const defaultState = {
@@ -60,10 +62,59 @@ const defaultState = {
   is_error: null,
   status_pct: null,
   needReload: false,
+  previousQuery: null,
 };
 
+function responseToQuery(response: DocTasksResponse): DocTasksQuery {
+  return {
+    doc_task_id: response.doc_task_id,
+    gvdbs_id: response.gvdbs_id,
+    gvdbs_cfg_json: JSON.parse(response.gvdbs_cfg_json),
+    gllms_id: response.gllms_id,
+    gc_id: response.gc_id,
+    short_name: response.short_name,
+    input_text: response.input_text,
+    optional_text: response.optional_text,
+  };
+}
+
+function compareWithPreviousQuery(
+  query1: DocTasksQuery,
+  query2: DocTasksQuery | null
+): boolean {
+  function trimStr(str: string) {
+    // convert multiple whitespaces to single space + trim
+    return str.replace(/\s+/g, " ").trim();
+  }
+  if (!(query1 && query2)) return false;
+  if (query1.doc_task_id !== query2.doc_task_id) return false;
+  if (query1.gvdbs_id !== query2.gvdbs_id) return false;
+  if (query1.gllms_id !== query2.gllms_id) return false;
+  if (query1.gc_id !== query2.gc_id) return false;
+  if (trimStr(query1.input_text) !== trimStr(query2.input_text)) return false;
+  if (trimStr(query1.optional_text) !== trimStr(query2.optional_text))
+    return false;
+  const cfg1 = query1.gvdbs_cfg_json;
+  const cfg2 = query2.gvdbs_cfg_json;
+  if (!(cfg1 && cfg2)) return false;
+  const sk1 = cfg1.search_kwargs;
+  const sk2 = cfg2.search_kwargs;
+  if (!(sk1 && sk2)) return false;
+  if (cfg1.search_type !== cfg2.search_type) return false;
+  if (Number(sk1.k) !== Number(sk2.k)) return false;
+  if (cfg1.search_type === "mmr") {
+    if (Number(sk1.fetch_k) !== Number(sk2.fetch_k)) return false;
+    if (Number(sk1.lambda_mult) !== Number(sk2.lambda_mult)) return false;
+  }
+  if (cfg1.search_type === "similarity_score_threshold") {
+    if (Number(sk1.score_threshold) !== Number(sk2.score_threshold))
+      return false;
+  }
+  return true;
+}
+
 export const useDocTasksCurrentStore =
-  createResettableStore<DocTasksCurrentState>((set) => ({
+  createResettableStore<DocTasksCurrentState>((set, get) => ({
     ...defaultState,
     setGVDBsID: (gvdbs_id: number | null) => set({ gvdbs_id }),
     setGLLMsID: (gllms_id: number | null) => set({ gllms_id }),
@@ -71,13 +122,22 @@ export const useDocTasksCurrentStore =
     setShortName: (short_name: string) => set({ short_name }),
     setInputText: (input_text: string) => set({ input_text }),
     setOptionalText: (optional_text: string) => set({ optional_text }),
-    setBeforeServerResponse: (query: DocTasksQuery) => set(query),
-    setFromServerResponse: (response: DocTasksResponse) => set(response),
+    setBeforeServerResponse: (query: DocTasksQuery) => {
+      set({ ...query, previousQuery: query });
+    },
+    setFromServerResponse: (response: DocTasksResponse) => {
+      const previousQuery = get().previousQuery || responseToQuery(response);
+      set({ ...response, previousQuery });
+    },
     setNeedReload: (needReload: boolean) => set({ needReload }),
     setFromHistory: (item: DocTasksShortItem) =>
       set({
         doc_task_id: item.doc_task_id,
+        previousQuery: null,
         needReload: true,
       }),
     setNewQuery: () => set(defaultState),
+    isSameAsPreviousQuery: (query: DocTasksQuery) => {
+      return compareWithPreviousQuery(query, get().previousQuery);
+    },
   }));